fix the code as review required
